+++ init.rb
@@ -1,5 +1,4 @@
 require_plugin 'acts_as_scribe'
-require_plugin 'acts_as_taggable_on_steroids'
 require_plugin 'acts_as_abusable'
 require_plugin 'viking'