+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Edge
 * Installation template now asks about installing gems as sudo
 * new generator for creating plugin's migrations, on installation and after updates
 * using acts_as_rateable and acts_as_shareable as plugins, which fixes problem with timestamps while installing
+* using acts_as_taggable_on instead of acts_as_taggable_on_steroids_
+* comments can be posted to twitter (kudos to boonious)
 
 0.5.4
 ----